+++ ChangeLog.md
@@ -1,4 +1,8 @@
-### v0.8.0
+### [v0.9.0](https://github.com/kaoskorobase/hsc3-process/tree/v0.9.0)
+
+* Add `hostname` parameter to `Sound.SC3.Server.Process.withTransport`
+
+### [v0.8.0](https://github.com/kaoskorobase/hsc3-process/tree/v0.8.0)
 
 * Upgrade to `hosc 0.13` and `hsc3 0.13`
 * Use `Sound.SC3.Server.Enum` data types for NRT file format
@@ -11,7 +15,7 @@
 * Split internal server module `Sound.SC3.Server.Process.Internal` into separate package `hsc3-process-internal`
 * Module reorganisation
 
-### v0.4.0
+### [v0.4.0](https://github.com/kaoskorobase/hsc3-process/tree/v0.4.0)
 
 * Add support for **internal server transports** using [bindings-sc3](http://space.k-hornz.de/software/bindings-sc3)
 * Overhaul [ConfigFile](http://hackage.haskell.org/package/ConfigFile) interface